simplify login && reorganize files
