--- db/config/db_config.toml
@@ -1,2 +1,0 @@
-[database]
-url = "postgres://maplestory:pass@localhost:5432/maplestory"
