+++ db/src/lib.rs
@@ -1,20 +1,17 @@
 extern crate serde;
-
 #[macro_use]
 extern crate diesel;
-
 #[macro_use]
 extern crate serde_derive;
 
 use diesel::pg::PgConnection;
 use diesel::prelude::*;
+use settings::Settings;
 
-pub mod accounts;
-pub mod models;
 pub mod schema;
 mod settings;
 
-use settings::Settings;
+pub mod account;
 
 pub fn establish_connection() -> PgConnection {
     let database_url = Settings::new().unwrap().database.url;
