+++ net/src/packet/handle/login/char/char_list.rs
