+++ net/src/packet/handle/login/char/mod.rs
@@ -1,0 +1,2 @@
+pub mod char_list;
+pub mod check_name;
