+++ net/src/packet/handle/login/main/set_gender.rs
@@ -1,0 +1,32 @@
+use crate::{error::NetworkError, io::client::MapleClient, packet::handle::PacketHandler};
+use packet::Packet;
+
+pub struct SetGenderHandler {}
+
+impl SetGenderHandler {
+    pub fn new() -> Self {
+        SetGenderHandler {}
+    }
+
+    fn accept_logon(&self, _client: &mut MapleClient) -> Result<(), NetworkError> {
+        // println!("Logging in!");
+
+        // let mut return_packet = build::login::status::build_successful_login_packet();
+        // match client.send(&mut return_packet) {
+        //     Ok(_) => {
+        //         println!("Logon success packet sent.");
+        //         Ok(())
+        //     }
+        //     Err(e) => Err(NetworkError::CouldNotSend(e)),
+        // }
+        Ok(())
+    }
+}
+
+impl PacketHandler for SetGenderHandler {
+    fn handle(&self, _packet: &mut Packet, client: &mut MapleClient) -> Result<(), NetworkError> {
+        // TODO: Can implement this when we've got a DB...
+        println!("Ignoring Set Gender Packet.");
+        Ok(self.accept_logon(client)?)
+    }
+}
