+++ net/src/packet/handle/login/mod.rs
@@ -1,23 +1,15 @@
-mod accept_tos;
-mod after_login;
-mod char_list;
-mod check_name;
-mod guest_login;
-mod login_credentials_handler;
-mod login_start_handler;
-mod register_pin;
-mod server_status;
-mod set_gender;
-mod world_list;
+mod char;
+mod main;
+mod world;
 
-pub use self::accept_tos::AcceptTOSHandler;
-pub use self::after_login::AfterLoginHandler;
-pub use self::char_list::CharListHandler;
-pub use self::check_name::CheckCharNameHandler;
-pub use self::guest_login::GuestLoginHandler;
-pub use self::login_credentials_handler::LoginCredentialsHandler;
-pub use self::login_start_handler::LoginStartHandler;
-pub use self::register_pin::RegisterPinHandler;
-pub use self::server_status::ServerStatusHandler;
-pub use self::set_gender::SetGenderHandler;
-pub use self::world_list::WorldListHandler;
+pub use self::char::char_list::CharListHandler;
+pub use self::char::check_name::CheckCharNameHandler;
+pub use self::main::accept_tos::AcceptTOSHandler;
+pub use self::main::after_login::AfterLoginHandler;
+pub use self::main::guest_login::GuestLoginHandler;
+pub use self::main::login::LoginCredentialsHandler;
+pub use self::main::login_start::LoginStartHandler;
+pub use self::main::register_pin::RegisterPinHandler;
+pub use self::main::set_gender::SetGenderHandler;
+pub use self::world::server_status::ServerStatusHandler;
+pub use self::world::world_list::WorldListHandler;
