+++ net/src/packet/handle/mod.rs
@@ -22,28 +22,24 @@ pub fn get_handler(op: i16) -> Box<dyn PacketHandler> {
     match num::FromPrimitive::from_i16(op) {
         Some(RecvOpcode::LoginCredentials) => Box::new(login::LoginCredentialsHandler::new()),
         Some(RecvOpcode::GuestLogin) => Box::new(login::GuestLoginHandler::new()),
+
         Some(RecvOpcode::ServerListReRequest) => Box::new(login::WorldListHandler::new()),
         Some(RecvOpcode::CharListRequest) => Box::new(login::CharListHandler::new()),
         Some(RecvOpcode::ServerStatusRequest) => Box::new(login::ServerStatusHandler::new()),
 
-        // TODO
-        Some(RecvOpcode::AcceptTOS) => Box::new(default::DefaultHandler::new()),
-        Some(RecvOpcode::SetGender) => Box::new(default::DefaultHandler::new()),
-
+        Some(RecvOpcode::AcceptTOS) => Box::new(login::AcceptTOSHandler::new()),
+        Some(RecvOpcode::SetGender) => Box::new(login::SetGenderHandler::new()),
         Some(RecvOpcode::AfterLogin) => Box::new(login::AfterLoginHandler::new()),
         Some(RecvOpcode::RegisterPin) => Box::new(login::RegisterPinHandler::new()),
         Some(RecvOpcode::ServerListRequest) => Box::new(login::WorldListHandler::new()),
 
-        // TODO
         Some(RecvOpcode::ViewAllChar) => Box::new(default::DefaultHandler::new()),
         Some(RecvOpcode::PickAllChar) => Box::new(default::DefaultHandler::new()),
         Some(RecvOpcode::CharSelect) => Box::new(default::DefaultHandler::new()),
-
         Some(RecvOpcode::CheckCharName) => Box::new(login::CheckCharNameHandler::new()),
-
-        // TODO
         Some(RecvOpcode::CreateChar) => Box::new(default::DefaultHandler::new()),
         Some(RecvOpcode::DeleteChar) => Box::new(default::DefaultHandler::new()),
+
         Some(RecvOpcode::RegisterPic) => Box::new(default::DefaultHandler::new()),
         Some(RecvOpcode::CharSelectWithPic) => Box::new(default::DefaultHandler::new()),
         Some(RecvOpcode::ViewAllPicRegister) => Box::new(default::DefaultHandler::new()),
